엔티티 setter없앤버전

+++ wooden/src/main/java/com/springboot/wooden/dto/PartRequestDto.java
@@ -2,8 +2,6 @@ package com.springboot.wooden.dto;
 
 import lombok.*;
 
-import java.math.BigDecimal;
-
 @Getter
 @Setter
 @NoArgsConstructor
@@ -15,7 +13,7 @@ public class PartRequestDto {
     private String partCode;
     private String partName;
     private String partSpec;
-    private BigDecimal partPrice;
+    private int partPrice;
 
     private Long buyerNo; // 등록/수정 시 FK로 받음
 }
